ensure detached before trying an attach
If an attach was partially successful, then it was a failure. Prior to
this change, subsequent attempts to attach (including retries) would
all fail because the machine was "already attached", even though the
original attach was partial (i.e. not all services were successfully
enabled). Now, we detach, disabling any services that got enabled in
a partial attach, before trying to attach again.

Fixes: #21